// web-next/src/api/schema.ts
export interface components {
    schemas: {
// ...
        SportAutoSignEntry: {
            /** Format: date-time */
            cancelledAt: string | null;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            expiredAt: string | null;
            /** Format: date-time */
            firstNotifiedAt: string | null;
            /** Format: int64 */
            id: number;
            isCancelled: boolean;
            /** Format: date-time */
            lastNotifiedAt: string | null;
            /** Format: int32 */
            maxNotificationAttempts: number;
            /** Format: int32 */
            notificationAttempts: number;
            /** Format: int32 */
            position: number;
            /** Format: int64 */
            prototypeLessonId: number;
            realLesson: components["schemas"]["SportLessonDto"] | null;
            /** Format: int64 */
            realLessonId: number | null;
            /** Format: date-time */
            satisfiedAt: string | null;
            /** @enum {string} */
            status: "WAITING" | "NOTIFIED" | "GAVE_UP_NOTIFYING" | "SATISFIED" | "EXPIRED";
            targetLesson: components["schemas"]["SportLessonDto"];
            /** Format: int32 */
            total: number;
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "auto";
        };
// ...
        SportFreeSignEntry: {
            /** Format: date-time */
            cancelledAt: string | null;
            /** Format: date-time */
            createdAt: string;
            /** Format: date-time */
            expiredAt: string | null;
            /** Format: date-time */
            firstNotifiedAt: string | null;
            forceSign: boolean;
            /** Format: int64 */
            id: number;
            isCancelled: boolean;
            /** Format: date-time */
            lastNotifiedAt: string | null;
            /** Format: int64 */
            lessonId: number;
            /** Format: int32 */
            maxNotificationAttempts: number;
            /** Format: int32 */
            notificationAttempts: number;
            /** Format: int32 */
            position: number;
            /** Format: date-time */
            satisfiedAt: string | null;
            /** @enum {string} */
            status: "WAITING" | "NOTIFIED" | "GAVE_UP_NOTIFYING" | "SATISFIED" | "EXPIRED";
            targetLesson: components["schemas"]["SportLessonDto"];
            /** Format: int32 */
            total: number;
            /**
             * @description discriminator enum property added by openapi-typescript
             * @enum {string}
             */
            type: "free";
        };
// ...
        SportLessonDto: {
            /** Format: int64 */
            buildingId: number | null;
            /** Format: date-time */
            end: string;
            /** Format: int64 */
            id: number;
            /** Format: int64 */
            level: number;
            roomName: string;
            /** Format: int64 */
            sectionId: number;
            /** Format: int64 */
            sectionLevel: number;
            sectionName: string;
            /** Format: date-time */
            start: string;
            teacherFio: string;
            /** Format: int64 */
            teacherIsu: number;
            /** Format: int64 */
            timeSlotId: number;
            /** Format: int64 */
            typeId: number;
        };
// ...
        UserCapabilities: {
            canViewFriends: boolean;
            canViewSchedule: boolean;
            canViewSport: boolean;
        };
// ...
        UserPrivacySettings: {
            /** @enum {string} */
            friendsVisibility: "ALL" | "FRIENDS" | "NOBODY";
            /** @enum {string} */
            scheduleVisibility: "ALL" | "FRIENDS" | "NOBODY";
            /** @enum {string} */
            sportVisibility: "ALL" | "FRIENDS" | "NOBODY";
        };
        UserProfile: {
            /** @enum {string} */
            relationship: "NONE" | "OUTGOING" | "INCOMING" | "FRIENDS" | "BLOCKED";
            user: components["schemas"]["UserData"];
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}